# apps/backend/src/main/resources/api.yaml
...
openapi: 3.0.0
info:
  title: StoreMe API
  version: "1"
servers:
  - url: http://localhost:8091
    description: Main API
...
paths:
  /authentication/getToken:
    post:
      tags:
        - Jwt
      summary: Get token
      description: Get a JWT token for a username and password.
      operationId: getToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsDTO"
      responses:
        "200":
          description: Operation successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWTDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
  /authentication/validateToken:
    post:
      tags:
        - Jwt
      summary: Validate token
      description: Verify that a token can be used.
      operationId: validateToken
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
      responses:
        "204":
          description: If the JWT is valid.
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
  "/inventory/{inventoryId}/categories":
    parameters:
      - in: path
        name: inventoryId
        schema:
          type: integer
        required: true
        description: Numeric of the current inventory
    get:
      tags:
        - Category
      description: Get categories
      operationId: getCategories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: [ ]
    post:
      tags:
        - Category
      description: Create category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryDTO"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: [ ]
  "/categories/{id}":
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the resource to get
    get:
      tags:
        - Category
      description: Get category
      operationId: getCategory
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
    put:
      tags:
        - Category
      description: Update category
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryDTO"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
    delete:
      tags:
        - Category
      description: Delete category
      operationId: deleteCategory
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
...
  /inventories:
    get:
      tags:
        - Inventory
      description: Get inventories, supports the swagger-undocumented filter
      operationId: getInventories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
    post:
      tags:
        - Inventory
      description: Create inventory
      operationId: createInventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryDTO"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
  "/inventories/{id}":
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the resource to get
    get:
      tags:
        - Inventory
      description: Get inventory
      operationId: getInventory
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
    put:
      tags:
        - Inventory
      description: Update inventory
      operationId: updateInventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryDTO"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
    delete:
      tags:
        - Inventory
      description: Delete inventory
      operationId: deleteInventory
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryDTO"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
      security:
        - JWTAuth: []
...
components:
  schemas:
    JSONErrorMessage:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
    CategoryDTO:
      x-implements:
        - ch.heig.pdg.backend.dto.IDataTransferObject<ch.heig.pdg.backend.entities.Category>
      required:
        - description
        - name
      properties:
        id:
          type: integer
        updated:
          type: string
        created:
          type: string
        name:
          type: string
        parentCategoryId:
          type: integer
          nullable: true
...
    InventoryDTO:
      x-implements:
        - ch.heig.pdg.backend.dto.IDataTransferObject<ch.heig.pdg.backend.entities.Inventory>
      required:
        - name
      properties:
        id:
          type: integer
        updated:
          type: string
        created:
          type: string
        name:
          type: string
        owner_id:
          type: integer
        users:
          type: array
          items:
            type: integer
    JWTDTO:
      type: object
      properties:
        token:
          type: string
    CredentialsDTO:
      required:
        - password
        - username
      type: object
      properties:
        username:
          type: string
        password:
          type: string
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"
    "401":
      description: No authorization was provided or token is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"
    "403":
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"
    "409":
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"
    "500":
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/JSONErrorMessage"